requirements.txt
django 
djangorestframework 
sklearn 
numpy